added 3 colors to the gradient

+++ apps/painterly_rendering.py
@@ -104,7 +104,7 @@ def main(args):
                                  stroke_width=torch.tensor(1.0),
                                  is_closed=True)
             shapes.append(path)
-            gradient = pydiffvg.LinearGradient(begin=torch.tensor([random.random()*canvas_width, random.random()*canvas_height]), end=torch.tensor([random.random()*canvas_width, random.random()*canvas_height]), offsets=torch.tensor([0.1, 0.5, 1],dtype=torch.float16), stop_colors=torch.tensor([[random.random(),
+            gradient = pydiffvg.LinearGradient(begin=torch.tensor([random.random()*canvas_width, random.random()*canvas_height]), end=torch.tensor([random.random()*canvas_width, random.random()*canvas_height]), offsets=torch.tensor([0.1, 0.5, 1],dtype=torch.float64), stop_colors=torch.tensor([[random.random(),
                                                                                                                                                                                                                                                                                    random.random(),
                                                                                                                                                                                                                                                                                    random.random(),
                                                                                                                                                                                                                                                                                    random.random()],
